Update decorator namespaces to be compliant with Zeitwerk
