+++ app/decorators/models/solidus_subscriptions/spree/order/after_create.rb
@@ -1,0 +1,17 @@
+module SolidusSubscriptions
+  module Spree
+    module Order
+      module AfterCreate
+        def ensure_line_items_present
+          super unless subscription_order?
+        end
+
+        def send_cancel_email
+          super unless subscription_order?
+        end
+      end
+    end
+  end
+end
+
+Spree::Order.prepend(SolidusSubscriptions::Spree::Order::AfterCreate)
