+++ app/models/solidus_subscriptions/installment_detail.rb
@@ -8,7 +8,7 @@ module SolidusSubscriptions
       inverse_of: :details
     )
 
-    belongs_to(:order, class_name: 'Spree::Order', optional: true)
+    belongs_to(:order, class_name: '::Spree::Order', optional: true)
 
     validates :installment, presence: true
     alias_attribute :successful, :success
