+++ app/models/solidus_subscriptions/subscription.rb
@@ -7,11 +7,11 @@ module SolidusSubscriptions
 
     PROCESSING_STATES = [:pending, :failed, :success]
 
-    belongs_to :user, class_name: Spree.user_class.to_s
+    belongs_to :user, class_name: "::#{::Spree.user_class}"
     has_many :line_items, class_name: 'SolidusSubscriptions::LineItem', inverse_of: :subscription
     has_many :installments, class_name: 'SolidusSubscriptions::Installment'
-    belongs_to :store, class_name: 'Spree::Store'
-    belongs_to :shipping_address, class_name: 'Spree::Address', optional: true
+    belongs_to :store, class_name: '::Spree::Store'
+    belongs_to :shipping_address, class_name: '::Spree::Address', optional: true
 
     validates :user, presence: :true
     validates :skip_count, :successive_skip_count, presence: true, numericality: { greater_than_or_equal_to: 0 }
