+++ app/models/solidus_subscriptions/subscription_line_item_builder.rb
@@ -8,8 +8,8 @@ module SolidusSubscriptions
       )
 
       # Rerun the promotion handler to pickup subscription promotions
-      Spree::PromotionHandler::Cart.new(line_item.order).activate
-      if Spree.solidus_gem_version >= Gem::Version.new('2.4.0')
+      ::Spree::PromotionHandler::Cart.new(line_item.order).activate
+      if ::Spree.solidus_gem_version >= Gem::Version.new('2.4.0')
         line_item.order.recalculate
       else
         line_item.order.update!
