+++ spec/decorators/models/solidus_subscriptions/spree/order/subscription_line_items_association_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+RSpec.describe SolidusSubscriptions::Spree::Order::SubscriptionLineItemsAssociation, type: :model do
+  subject { Spree::Order.new }
+
+  it { is_expected.to have_many :subscription_line_items }
+end
