+++ spec/decorators/models/solidus_subscriptions/spree/user/have_many_subscriptions.rb
@@ -1,0 +1,8 @@
+require 'spec_helper'
+
+RSpec.describe SolidusSubscriptions::Spree::User::HaveManySubscritptions, type: :model do
+  subject { Spree::User.new }
+
+  it { is_expected.to have_many :subscriptions }
+  it { is_expected.to accept_nested_attributes_for :subscriptions }
+end
