--- spec/overrides/spree/orders/subscription_line_items_association_spec.rb
@@ -1,7 +1,0 @@
-require 'spec_helper'
-
-RSpec.describe Spree::Orders::SubscriptionLineItemsAssociation, type: :model do
-  subject { Spree::Order.new }
-
-  it { is_expected.to have_many :subscription_line_items }
-end
