--- spec/overrides/spree/users/have_many_subscriptions.rb
@@ -1,8 +1,0 @@
-require 'spec_helper'
-
-RSpec.describe Spree::Users::HaveManySubscritptions, type: :model do
-  subject { Spree::User.new }
-
-  it { is_expected.to have_many :subscriptions }
-  it { is_expected.to accept_nested_attributes_for :subscriptions }
-end
